RegexpCompletor extends BaseCompletor, move back definition to completion.rb
